fastapi/auth_user_article/requirements.txt
anyio==3.6.2
asgiref==3.6.0
asyncpg==0.25.0
cffi==1.15.1
click==8.1.3
cryptography==39.0.0
dnspython==2.3.0
ecdsa==0.18.0
email-validator==1.3.1
fastapi==0.75.2
greenlet==2.0.1
h11==0.14.0
idna==3.4
passlib==1.7.4
psycopg2-binary==2.9.3
pyasn1==0.4.8
pycparser==2.21
pydantic==1.10.4
python-jose==3.3.0
python-multipart==0.0.5
pytz==2022.1
rsa==4.9
six==1.16.0
sniffio==1.3.0
SQLAlchemy==1.4.36
starlette==0.17.1
typing_extensions==4.4.0
uvicorn==0.17.6
